Unified build flag naming across main project and subprojects:
- Removed TINYCLJ_ prefix from build flags (ZOMBIE_ENABLED, META_ENABLED, MEMORY_PROFILING_ENABLED)
- Flags now propagate consistently from main CMakeLists.txt to subjective-c
- Updated documentation to reflect new flag names

This improves consistency and makes the build system more maintainable.

+++ src/line_editor.c
@@ -868,7 +868,7 @@ int line_editor_get_state(const LineEditor *editor, LineEditorState *state) {
     strncpy(state->buffer, src, sizeof(state->buffer) - 1);
     state->buffer[sizeof(state->buffer) - 1] = '\0';
     state->cursor_pos = editor->cursor_pos;
-    state->length = (int)editor->buffer.length;
+    state->length = editor->buffer.length;
     state->line_ready = editor->line_ready;
     
     return LINE_EDITOR_SUCCESS;
